+++ libs/my_c/src/string/my_strcapitalize.c
@@ -22,7 +22,7 @@
 char *my_strcapitalize(char *str)
 {
     str = my_strlowcase(str);
-    for (int i = 1; str[i] != '\0'; i++) {
+    for (size_t i = 1; str[i] != '\0'; i++) {
         if (((str[i - 1] == '+' || str[i - 1] == '-' || str[i - 1] == ' '))
         && (str[i] >= 'a' && str[i] <= 'z'))
             str[i] = str[i] - 32;
