+++ libs/my_c/src/string/my_strncpy.c
@@ -22,11 +22,11 @@
  * @return <b>char*</b> The destination string
  * @author Nicolas TORO
  */
-char *my_strncpy(char *dest, char const *src, int n)
+char *my_strncpy(char *dest, char const *src, size_t n)
 {
-    int len = my_strlen(src);
+    size_t len = my_strlen(src);
 
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         dest[i] = src[i];
     if (len < n)
         dest[len] = '\0';
